requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
transformers
huggingface-hub
requests
sentence-transformers
numpy
python-docx
python-dotenv